+++ src/Hacks/FrameStepper.cpp
@@ -1,0 +1,6 @@
+#include <Geode/Geode.hpp>
+#include <Geode/modify/PlayLayer.hpp>
+#include "../Client/Client.h"
+
+using namespace geode::prelude;
+